# Arrow2 cherry pick and improve hashing performance