--- a/polars/polars-arrow/src/trusted_len/mod.rs
+++ b/polars/polars-arrow/src/trusted_len/mod.rs
@@ -3,6 +3,7 @@ mod boolean;
 use crate::utils::{FromTrustedLenIterator, TrustMyLength};
 use arrow::bitmap::utils::{BitmapIter, ZipValidity};
 use arrow::buffer::MutableBuffer;
+use arrow::types::NativeType;
 use std::slice::Iter;
 
 /// An iterator of known, fixed size.
@@ -81,6 +82,34 @@ pub trait PushUnchecked<T> {
 
     /// Extend the array with an iterator who's length can be trusted
     fn extend_trusted_len<I: IntoIterator<Item = T> + TrustedLen>(&mut self, iter: I);
+}
+
+impl<T: NativeType> PushUnchecked<T> for MutableBuffer<T> {
+    unsafe fn push_unchecked(&mut self, value: T) {
+        let end = self.as_mut_ptr().add(self.len());
+        std::ptr::write(end, value);
+        self.set_len(self.len() + 1);
+    }
+
+    unsafe fn push_unchecked_no_len_set(&mut self, value: T) {
+        let end = self.as_mut_ptr().add(self.len());
+        std::ptr::write(end, value);
+    }
+
+    fn extend_trusted_len<I: IntoIterator<Item = T> + TrustedLen>(&mut self, iter: I) {
+        let iter = iter.into_iter();
+        let upper = iter.size_hint().1.expect("must have an upper bound");
+        self.reserve(upper);
+
+        unsafe {
+            let mut dst = self.as_mut_ptr().add(self.len());
+            for value in iter {
+                std::ptr::write(dst, value);
+                dst = dst.add(1)
+            }
+            self.set_len(self.len() + upper)
+        }
+    }
 }
 
 impl<T> PushUnchecked<T> for Vec<T> {