--- a/py-polars/src/conversion.rs
+++ b/py-polars/src/conversion.rs
@@ -4,6 +4,7 @@ use crate::prelude::*;
 use crate::series::PySeries;
 use polars::chunked_array::object::PolarsObjectSafe;
 use polars::frame::row::Row;
+use polars::frame::NullStrategy;
 use polars::prelude::AnyValue;
 use polars_core::utils::arrow::types::NativeType;
 use pyo3::basic::CompareOp;
@@ -14,7 +15,6 @@ use pyo3::types::PySequence;
 use pyo3::{PyAny, PyResult};
 use std::fmt::{Display, Formatter};
 use std::hash::{Hash, Hasher};
-use polars::frame::NullStrategy;
 
 #[repr(transparent)]
 pub struct Wrap<T>(pub T);