Improves hashing of data with and without nulls for numeric columns.


# vec_hash

```
bench vec hash null rate: 0%                                                                            
                        time:   [71.386 us 71.394 us 71.403 us]
                        change: [-10.711% -10.525% -10.359%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 27 outliers among 100 measurements (27.00%)
  20 (20.00%) low severe
  1 (1.00%) low mild
  1 (1.00%) high mild
  5 (5.00%) high severe

bench vec hash null rate: 20%                                                                            
                        time:   [168.02 us 168.23 us 168.46 us]
                        change: [-21.523% -21.441% -21.336%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  7 (7.00%) high mild
  1 (1.00%) high severe

bench vec hash null rate: 50%                                                                            
                        time:   [247.60 us 247.68 us 247.78 us]
                        change: [-16.209% -15.960% -15.762%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  1 (1.00%) high mild
  6 (6.00%) high severe

bench vec hash null rate: 80%                                                                            
                        time:   [173.88 us 174.69 us 175.69 us]
                        change: [-8.4531% -7.7813% -6.7419%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high severe
```

# vec_hash_combine

```
bench vec hash null rate: 0%                                                                            
                        time:   [86.693 us 86.774 us 86.865 us]
                        change: [-35.625% -35.426% -35.189%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 18 outliers among 100 measurements (18.00%)
  6 (6.00%) low severe
  5 (5.00%) high mild
  7 (7.00%) high severe

bench vec hash null rate: 20%                                                                            
                        time:   [219.04 us 219.39 us 219.77 us]
                        change: [-12.193% -12.008% -11.803%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 11 outliers among 100 measurements (11.00%)
  8 (8.00%) high mild
  3 (3.00%) high severe

bench vec hash null rate: 50%                                                                            
                        time:   [291.01 us 291.09 us 291.16 us]
                        change: [-15.787% -15.603% -15.438%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild

bench vec hash null rate: 80%                                                                            
                        time:   [176.61 us 176.66 us 176.70 us]
                        change: [-25.989% -25.894% -25.798%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) low mild
  2 (2.00%) high mild
  ```