@ritchie46 2021-08-29
@jorgecarleitao FYI. This branch actually used the iterator provide by arrow2.